Add confirm-create-channel-workflow component
able to send events by clicking buttons
application workflow component split out into seperate file
top level wallet component switches on machine id
